Add an empty `.watchmanconfig` file. (#697)
Summary:
**Summary**
When running Metro from source (or when running Metro's tests with Jest!) it's good to have an empty `.watchmanconfig` file. This is primarily necessary for non-Facebook issued laptops that don't have some magic setup related to watchman and chef. Jest [does the same thing](https://github.com/facebook/jest/blob/master/.watchmanconfig).

**Test plan**
* Run Metro from source on a non-FB issued laptop that uses watchman (See https://github.com/facebook/metro/issues/696).
* It does not cause trouble.

Note: you may not be able to merge this PR via Facebook's GitHub plugins and "shipit" because this file may be ignored. In this case, log in to Safari and merge it directly into the repo.

Pull Request resolved: https://github.com/facebook/metro/pull/697

Reviewed By: GijsWeterings

Differential Revision: D30129414

Pulled By: motiz88

fbshipit-source-id: 4cccd864cb423a2152c25675249992c09baf8e9d