Support package.json module replacements for requires with implicit extensions
Summary: When [replacing modules in `browser`-like-fields in `package.json`](https://github.com/defunctzombie/package-browser-field-spec#replace-specific-files---advanced), if a replacement is specified as `./foo.js`, Metro currently does not apply the replacement to `require('./foo')`. Here we change the resolver to correctly look for all extension variants of a path in `package.json`.

Reviewed By: GijsWeterings

Differential Revision: D30776378

fbshipit-source-id: 2bca08bbee6bfa731bcd1cbd75a522fc9440bfa3