Store graph entry points as a Set
Summary: I'm working on a change to `traverseDependencies()` that requires an O(1) check of whether a given module path is an entry point of the graph. Here I'm changing `DeltaBundler` et al to store the entry points as a `Set` (without affecting DeltaBundler's API). Entry points are deduplicated by the traversal logic so this is functionally a no-op.

Reviewed By: GijsWeterings

Differential Revision: D36403389

fbshipit-source-id: 272dc4b970d7960b43788102de6dc0a5b9c49741