Added basic Jest test for the metro-react-native-interop-tools
Summary:
We're adding a unit test that tests the functionality of the metro-react-interop-tools entrypoint.
This helps us validate the output of the package's run function so as we build functionality, we can continuously verify that the package outputs what we're expecting in an automated flow, so any future change to this package can be protected from unexpected output regression.

Reviewed By: GijsWeterings

Differential Revision: D37818310

fbshipit-source-id: 81c244de96be202c8057b6dffe09f0758d744597