metro-react-native-interop-tools added getTypeAnnotation function
Summary:
getTypeAnnotation gets a TypeAnnotation AST node and returns the type of the node.
This function removes other properties from the ast node, which are not helpful in creating the type schema.

Most parts of the type schema are the type annotations, getTypeAnnotation will be called from another function which creates the main structure of our schema. Adding the type annotations will complete the entire type schema.

There are some missing types for now. The more complex types(object, function, array, etc.) need to be handled differently to be easier to understand.

Reviewed By: GijsWeterings

Differential Revision: D38144350

fbshipit-source-id: 6454f6edbe8e11b961d3c859177e376117c982c3