metro-react-native-interop-tools added type diffing
Summary:
https://pxl.cl/2bfrV
After we got the boundary schema from the AST we are using that schema in version diffing, which takes two boundary schemas and compares if there are incompatibilities between them.
The compareTypeAnnotation compares only the annotation types between boundary schemas, It will take the body of our interface, which is an ObjectTypeAnnotation and recursively compare each type.
I added just the basic annotation types, but the complex types(object, function, etc.) will have helpers to be easier to understand.

One important thing here is the result that I’m returning after comparing the types. The function returns an array of strings which are the incompatibilities that will be shown to the developer. The makeError function is formatting the incompatibility to be readable by the developer. In the next diffs, I will also add the location of the code that's creating the incompatibility.

The default case is unknown: the type is not handled yet by the comparator, when I will add all possible types I will change the return with throwing an error which shows that the tool has a problem.
[1]
```
export interface Spec extends TurboModule {
   +setKeepScreenOn: () => void;
}
```
For the code above this is the boundary schema:
```
{
 "typegenSchema": {
   "Spec": {
     "typeAnnotation": {
       "type": "InterfaceDeclarationTypeAnnotation",
       "body": {
         "type": "ObjectTypeAnnotation",
         "loc": {...},
         "properties": [
           {
             "loc": {...},
             "name": "setKeepScreenOn",
             "optional": false,
             "typeAnnotation": {
               "type": "FunctionTypeAnnotation",
               "loc": {...},
               "params": [
                 {
                   "name": "screenShouldBeKeptOn",
                   "typeAnnotation": {
                     "type": "BooleanTypeAnnotation",
                     "loc": {...}
                   }
                 }
               ],
               "returnTypeAnnotation": {
                 "type": "VoidTypeAnnotation",
                 "loc": {...}
                 }
               }
             }
           }
         ]
       },
       "extends": [
         {
           "loc": {
             "start": {...},
           "name": "TurboModule",
           "typeParameters": []
         }
       ]
     }
   }
 }
}
```

Reviewed By: motiz88

Differential Revision: D38978074

fbshipit-source-id: 6bfae4447baae3dfae52e7f5afee22a928c5776f