Move experimentalImportBundleSupport option from transformer to server
Summary:
`experimentalImportBundleSupport` (introduced in D15943150 (https://github.com/facebook/metro/commit/72329d02f2513eec024c42aa777bd445f573a968)) is never used to change the transformer's output for a given module, so having it under `transformer` is (1) confusing and (2) inefficient, because it gets added to the transformer cache key and therefore forks the transformer cache. Here we move it to `server` while keeping it as an experimental, undocumented option (for now).

This change removes one of the blockers to officially releasing this feature in open source: `metro build` only ever makes sense with `experimentalImportBundleSupport: false`, but prior to this diff, it couldn't share a cache with `metro serve` running with `experimentalImportBundleSupport: true` - so enabling this feature in the server would have regressed Metro's overall perceived performance on a given machine.

Changelog: [Internal]

Reviewed By: jacdebug

Differential Revision: D40389568

fbshipit-source-id: 644d070ff069e8396b9f44e8429147a42727eb5f